Add Devise and minimal view support.
This commit adds Devise support.  There are loads of cool features you
can get with Devise, but it seems like the boilerplate works here (for
now).  We may have to eject the Devise views / controllers at a later
time if the application experience demands it, but for now we just use
the stock views.

* Add Devise gem.
* Creates a User model.
* Creates a standard Devise install.
* Put a small partial which links to Devise functionality in the app
  layout.